+++ NEWS.md
@@ -1,3 +1,5 @@
+# cheapr (development version)
+
 # cheapr 0.9.91
 
 * New functions `cheapr_if_else`, `case` and `val_match` to make