+++ NEWS.md
@@ -1,3 +1,5 @@
+# cheapr (development version)
+
 # cheapr 0.9.92
 
 - Fixed additional issues flagged by R checks.