+++ NEWS.md
@@ -2,6 +2,12 @@
 
 - New function `cheapr_table` to create fast frequency tables.
 
+- `count_val`, `unused_levels` and `used_levels` have been removed
+removed and replaced by `val_count`, `levels_unused` and `levels_used` 
+respectively.
+
+- Deprecated `cut_numeric`, `enframe_` and `deframe_`.
+
 # cheapr 0.9.92
 
 - Fixed additional issues flagged by R checks.